Add tags in a transaction.
In this case, the key can't expire directly after adding (minimum expiration is one second)
so it is safe to assume that it exists and that we should add tags for it.

We cannot do the same when using exipreat as we cannot rely on the redis server having the
exact same time as the server beatbox is running on.